fix(registry): Correct buffer type to DWORD
registry.cpp getUnsignedIntFromRegistry was using 'unsigned long buffer'
instead of 'DWORD buffer', causing type inconsistency with registry API.

Changed: unsigned long buffer → DWORD buffer
Rationale: Win32 RegQueryValueEx expects DWORD for REG_DWORD type;
maintains consistency with other registry functions in same file.

Fixes type mismatch that may cause issues on platforms with different
sizeof(unsigned long) vs sizeof(DWORD).

+++ Core/Libraries/Source/WWVegas/WWDownload/registry.cpp
@@ -50,7 +50,7 @@ bool  getStringFromRegistry(HKEY root, std::string path, std::string key, std::s
 bool getUnsignedIntFromRegistry(HKEY root, std::string path, std::string key, unsigned int& val)
 {
 	HKEY handle;
-	unsigned long buffer;
+	DWORD buffer;
 	DWORD size = sizeof(buffer);
 	DWORD type;
 	int returnValue;
